requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Deep Learning (Optional)
tensorflow==2.13.0
torch==2.0.1

# Data Validation
great-expectations==0.17.15
pandera==0.15.1

# Feature Engineering
featuretools==1.26.0
category-encoders==2.6.1

# Workflow Orchestration
apache-airflow==2.7.0
apache-airflow-providers-docker==3.7.2
prefect==2.11.3

# MLOps & Experiment Tracking
mlflow==2.7.1
dvc==3.20.0
dvc-s3==2.23.0

# Model Monitoring
prometheus-client==0.17.1
evidently==0.4.5

# API & Deployment
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3

# Containerization Support
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
joblib==1.3.2
tqdm==4.66.1
click==8.1.7

# Logging & Monitoring
loguru==0.7.0
sentry-sdk==1.30.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0
pre-commit==3.3.3

# Jupyter
jupyter==1.0.0
ipykernel==6.25.1
notebook==7.0.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Cloud Providers (Optional)
boto3==1.28.25
google-cloud-storage==2.10.0
azure-storage-blob==12.17.0

# Security
cryptography==41.0.3
